Add Delivery logic
I updated context.py, adding the delivery logic. The website
offers 2 types of delivery (standard and High Quality),
which can be selected by a user from a dropdown selector. Once the
delivery method is chosen, the quantity associated with the
selected method is summed with the subtotal, resulting in the
purchase total.