Update Delivery logic
I modified the delivery logic by fixed prices instead of calculating
by weight. The reason for doing this was that it was a very
cumbersome process for the user, the page didn't update immediately
every time the user changed the delivery type, and every time the user
added too many items, the delivery values didn't adjust or simply
made wrong calculations. To fix this, I added fixed values to
the 2 delivery types created before, and at the same time I added a
third one, which is free delivery, that is activated every time
the user spends more than 60 euros on the website.