Change Delivery logic
Due to compatibility problems between the delivery logic created in
contexts.py, and the classes created in my document
checkout/models.py, I had to change all the delivery logic and adjust
it to calculate the price of this, because when the user makes his/her
purchase and creates an order, the purchase ticket did not save the
prices correctly. In addition, for reasons related to time and the
complexity of trying to fix the problem,  I decided to change the
scope that I had thought for my delivery system, and I replaced it
with a system that calculates 10% of the total purchase as the price
of standard delivery, and when the user spends more than 60 euros,
the delivery is totally free.