Update Checkout forms
The update consists on that when the user is going to fill the
purchase form, instead of typing the country name now he/she can
select it from a dropdown box. To do this I installed a Django library
called Django Countries, which provides a list of countries to select
from. To make the project feasible and at the same time fit the
reality of it, I created a custom list of countries to which the
purchases will be delivered, these are: Republic of Ireland and
United Kingdom.