Add confirmation function to delete a product
I've added the functionality for ask the super user before deleting
a product, because before this action was done immediately, and there
could be cases of unintentional deletion of an item. This function has
been done in 2 modalities, in the product details template, if the
admin presses the delete button, a modal will appear asking if he/she
is sure to perform the action. In the product template, if the admin
presses the delete link, an alert message will appear asking if he/she
is sure to delete the selected item.