Update Home page
The update made for the Home page is that when starting the website,
in the Featured products and Combos section, these will render a
random amount of products. To do this I added 2 variables in my
function created in the views of this app, which are responsible for
filtering all products on the website, showing only a certain amount
of items.