Add editing user bio functionality
Website users are able to edit the bio within their created profile.
For this I implemented a button, which when pressed, will display a
modal window containing the form to edit it. I also updated the
profile views and forms, in order to incorporate this new function
to edit elements inside the user profile.