Update editing profile functionality
I have modified the profiles views in order to better adapt the
function of editing and updating the user's profile, since with the
previous implementation of editing the user's bio, this was updated
unsynchronized with the delivery information form. The main problem
that arose was that when updating a form, it deleted the information
inside the other, creating a big chaos, and at the same time it didn't
allow to add or update the user's avatar image. The solution I applied
was to modify the bio form and add the image field in it, and then I
added the image field inside the logic of the user's form with
information for deliveries. By making this modification, I managed to
solve the problem of unsynchronized update that these previously
presented, also when the user presses the button to edit the profile,
this will be able to edit the forms from a modal window, being more
comfortable and intuitive.