Update blog app
I added a new class inside the models.py file in order to allow users
to leave a comment on the posts published on the website. I have also
created a new form that allows to post comments, which will first be
moderated by the super user in order to avoid spams or comments with
questionable content.