Implement customer reviews functionality
I implemented the function of adding user reviews to the website. This
feature is available to users who have a profile created, and they can
do it by filling a form which contains the user's name, email, a field
to write their comment and a dropdown to select their rating (the
ranges are from 1 to 5 stars). Once the review is created, it will be
under moderation of the super user, as a defense mechanism to avoid
inappropriate comments or hate messages.