Update cart and products templates
In this update I've improved the shopping cart view when it's empty,
for this I added an image and adjusted the html elements of the page.
In the case of the products template, I added a continue shopping
button in case the user writes a word in the search bar and it
doesn't find anything related to what the user requested.