Extends and customizes django-allauth sign up form
I have added 2 new text fields for when a user signs up to the online
store.  These text fields include the following:

- First Name
- Last Name

I decided to make this customization in the sign up form, because
when the user was registering, those 2 fields were empty inside the
administration panel of the website.