Update checkout views
The update is that if the user is going to make a purchase, instead
of loading the data from default_name or default_email, it now does
it directly from the data entered by the user when creating a profile
account. This saves time for the first time the user makes a purchase.