Update blog and reviews app
I updated the views of both apps modifying the way to extract the user
data when leaving a comment or review on the website, in these form
fields will load the full name and email given at the moment of
registering in the online store instead of taking the default_name
and default_email fields. I also modified the review models in order
to show 1 star by default in the rating field of the form.