Add avatar image resizing functionality
I updated the profile templates in order to prevent users from
uploading heavy images to the database. For this I've created a
function that rescales the avatar image that the user wants to add
in the user's profile, and this operation happens before the image
is saved in the database.