Update contact form and views
I added placeholders and custom labels inside the contact form fields.
I've also implemented the functionality to pre-fill the full name and
email address fields inside this form using the data saved within the
user's profile, and this is only activated if the user is logged in
and actually has an account created.