Add Procfile
I added a Procfile file in order to deploy the website using Heroku.
For this I've updated the settings.py file in order to configure the
project and make it connect to the remote database. I also updated the
requirements.txt file with the new libraries installed to make
possible the deployment of the online store.