Add Features section and wireframes
I added the features section inside the README.md file in order to
specify all the features implemented in the web site, features that
couldn't be implemented as well as features that I would like to
implement in the future. I updated the wireframes, adding them all
in a PDF file.