Update of project images
I updated the images for the products and blogposts in order to adjust
them to a standard size regardless of the dimensions of the image
uploaded within the project. The function of this change is to ensure
that the website does not break down or stop looking harmonious in
case of uploading images with different dimensions than the standard
ones.