Update Manual testing section
Update the TESTING.md file in order to add the Issues during project
development section, which is the last section of the manual tests
performed on this project. In this section are mainly described some
of the situations that caused problems during the project development,
and each one has a solution given.